Fix timeval-to-timespec conversions (#814)
Time values returned by gettimeofday() have microseconds in
timeval::tv_usec while timespec::ts_nsec is nanoseconds.

timeval::tv_usec is 64 bit (at least on some platforms), while
timespec::tv_nsec is always 'long'. An explicit cast is required to
avoid -Wnarrowing warnings/errors.

+++ Pcap++/src/PcapFileDevice.cpp
@@ -188,7 +188,7 @@ bool PcapFileReaderDevice::getNextPacket(RawPacket& rawPacket)
 	uint8_t* pMyPacketData = new uint8_t[pkthdr.caplen];
 	memcpy(pMyPacketData, pPacketData, pkthdr.caplen);
 #if defined(PCAP_TSTAMP_PRECISION_NANO)
-	timespec ts = { pkthdr.ts.tv_sec, pkthdr.ts.tv_usec }; //because we opened with nano second precision 'tv_usec' is actually nanos
+	timespec ts = { pkthdr.ts.tv_sec, static_cast<long>(pkthdr.ts.tv_usec) }; //because we opened with nano second precision 'tv_usec' is actually nanos
 #else
 	struct timeval ts = pkthdr.ts;
 #endif
